+++ packages/tddy-web/src/components/GhosttyTerminal.stories.tsx
@@ -79,6 +79,9 @@ function LiveKitConnectedStory(args: {
       showBufferTextForTest={showBufferTextForTest}
       debugMode={debugMode}
       debugLogging={debugLogging}
+      connectionOverlay={{
+        onDisconnect: () => {},
+      }}
     />
   );
 }